[Refactor] port ClassUtils to HiveMetaClient
